+++ src/index.spec.js
@@ -1,4 +1,5 @@
 import { ChatSession } from "./index";
+import pkg from '../package.json';
 
 const ChatSessionObject = {
     csmService: "csmService"
@@ -24,5 +25,23 @@ describe("Chat JS index file", () => {
         expect(global.connect.LogManager).toEqual(LogManager);
         expect(global.connect.LogLevel).toEqual(LogLevel);
         expect(global.connect.csmService).toEqual("csmService");
+    });
+
+    test('should expose ChatJS.version with correct version', () => {
+        console.log(global.connect.ChatJS);
+        expect(global.connect.ChatJS.version).toBeDefined();
+        expect(global.connect.ChatJS.version).toBe(process.env.npm_package_version); // "3.x.x"
+        expect(global.connect.ChatJS.version).toBe(pkg.version); // "3.x.x"
+    });
+
+    test('should not allow modification of ChatJS.version', () => {
+        const originalValue = global.connect.ChatJS.version;
+
+        expect(() => {
+            global.connect.ChatJS.version = 'MODIFIED VERSION';
+        }).toThrow(TypeError);
+
+        expect(global.connect.ChatJS.version).toBe(originalValue);
+        expect(global.connect.ChatJS.version).not.toBe('MODIFIED VERSION');
     });
 });